# apps/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students/signup": {
      "get": {
        "operationId": "studentSignup",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/" }
            }
          }
        },
        "responses": { "200": { "description": "" } }
      }
    },
    "/students/login": {
      "post": {
        "operationId": "studentLogin",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StudentLoginDto" }
            }
          }
        },
        "responses": { "200": { "description": "student logged in" } }
      }
    },
...
    "/students/profile": {
      "get": {
        "operationId": "StudentsController_getProfile",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      },
      "post": {
        "operationId": "StudentsController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdateProfileDto"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      }
    },
    "/students/questions/{questionId}": {
      "get": {
        "operationId": "StudentsController_getQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    },
...
  },
  "info": {
    "title": "Questionnaire Open Api",
    "description": "Questionnaire api description",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "StudentLoginDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "StudentUpdateProfileDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "firstname", "lastname", "password"]
      },
...
    }
  }
}